移动mod domain提取逻辑
