+++ src/main/java/i18nupdatemod/launchwrapper/LaunchWrapperTweaker.java
@@ -2,12 +2,12 @@ package i18nupdatemod.launchwrapper;
 
 import i18nupdatemod.I18nUpdateMod;
 import i18nupdatemod.util.Log;
+import i18nupdatemod.util.ModUtil;
 import i18nupdatemod.util.Reflection;
 import net.minecraft.launchwrapper.ITweaker;
 import net.minecraft.launchwrapper.LaunchClassLoader;
 
 import java.io.File;
-import java.util.HashSet;
 import java.util.List;
 
 //1.6-1.12.2
@@ -21,7 +21,7 @@ public class LaunchWrapperTweaker implements ITweaker {
             Log.warning("Failed to get minecraft version.");
             return;
         }
-        I18nUpdateMod.init(gameDir.toPath(), mcVersion, "Forge", new HashSet<>());
+        I18nUpdateMod.init(gameDir.toPath(), mcVersion, "Forge", ModUtil.getModDomainsFromModsFolder(gameDir.toPath(), mcVersion, "Forge"));
     }
 
     @Override
